Move MeterProvider initialization to service/telemetry
