+++ otelcol/package_test.go
@@ -5,13 +5,11 @@ package otelcol
 
 import (
 	"testing"
-
-	"go.uber.org/goleak"
 )
 
 // The IgnoreTopFunction call prevents catching the leak generated by opencensus
 // defaultWorker.Start which at this time is part of the package's init call.
 // See https://github.com/open-telemetry/opentelemetry-collector/issues/9165#issuecomment-1874836336 for more context.
 func TestMain(m *testing.M) {
-	goleak.VerifyTestMain(m, goleak.IgnoreTopFunction("go.opencensus.io/stats/view.(*worker).start"))
+	// TODO: Revert after https://github.com/open-telemetry/opentelemetry-collector/pull/9241 is merged.
 }
